noir_ast_fuzzer
Module compare

Modules

compiled 🔒

comptime 🔒

Structs

CompareArtifact

CompareCompiled

Compare the execution of equivalent programs, compiled in different ways.

CompareComptime

Compare the execution of a Noir program in pure comptime (via interpreter) vs normal SSA execution.

CompareOptions

Subset of [SsaEvaluatorOptions] that we want to vary.

ExecOutput

Enums

CompareResult

Possible outcomes of the differential execution of two equivalent programs.

Traits

HasPrograms

Help iterate over the program(s) in the comparable artifact.

Type Aliases

CompareMorph

Compare two equivalent variants of the same program, compiled the same way.

ComparePipelines

Compare the execution the same program compiled in two different ways.

ExecResult 🔒