noir_ast_fuzzer::compare

Module compiled

Compare an arbitrary AST compiled into bytecode and executed with the VM.

Structs

CompareArtifact

CompareCompiled

Compare the execution of equivalent programs, compiled in different ways.

NargoErrorWithTypes

Type Aliases

CompareCompiledResult

The result of the execution of compiled programs, decoded by their ABI.

CompareMorph

Compare two equivalent variants of the same program, compiled the same way.

ComparePipelines

Compare the execution the same program compiled in two different ways.

ExecResult 🔒

The execution result is the value returned from the circuit and any output from println.

SsaErrorTypes 🔒

These are the error types in the SsaProgramArtifact, which are not the same as the ones in the ABI, but they can provide extra information when comparing errors.