noir_ast_fuzzer::compare
Module interpreted

Compare an arbitrary AST compiled into SSA and executed with the SSA interpreter at some stage of the SSA pipeline.

Structs

CompareInterpreted

Inputs for comparing the interpretation of two SSA states of an arbitrary program.

ComparePass

The state of the SSA after a particular pass in the pipeline.

Functions

append_input_type_to_ssa 🔒

append_input_value_to_ssa 🔒

input_type_to_ssa 🔒

Convert an ABI type into SSA.

input_value_to_ssa

Convert one ABI encoded input to what the SSA interpreter expects.

input_values_to_ssa

Convert the ABI encoded inputs to what the SSA interpreter expects.

sanitize_ssa 🔒

Remove identifiers from the SSA, so we can compare the structure without worrying about trivial differences like changing IDs of the same variable between one pass to the next.

Type Aliases

CompareInterpretedResult

The result of the SSA interpreter execution.

InterpretResult 🔒