noir_ast_fuzzer::compare

Module logging

We can turn on the logging of artifacts by setting the RUST_LOG=debug env var.

This can help reproducing failures. The functions in this module can help cut back some repetition. Log things as soon as they are available in case the next step fails.

Functions

format_msg 🔒

log_abi_inputs 🔒

log_comptime 🔒

log_options 🔒

log_program 🔒

log_ssa 🔒

log_ssa_inputs 🔒