noir_ast_fuzzer
Module program

Module responsible for generating arbitrary Program ASTs.

Modules

expr

freq 🔒

func 🔒

rewrite

scope

types

Structs

Context 🔒

Context to accumulate top level generated item, so we know what we can choose from.

DisplayAstAsNoir

Wrapper around Program that prints the AST as close to being able to copy-paste it as a Noir program as we can get.

DisplayAstAsNoirComptime

Wrapper around Program that prints its AST as close to Noir syntax as we can get, making it comptime. The AST must be specifically prepared to include a main function consisting of a comptime wrapped call to a comptime (or unconstrained) marked function.

Enums

CallableId 🔒

ID of a function we can call, either as a pointer in a local variable, or directly as a global function.

VariableId 🔒

ID of variables in scope.

Constants

CONSTRAIN_MSG_LENGTH 🔒

Length of generated random constraint messages.

Functions

arb_program

Generate an arbitrary monomorphized AST.

arb_program_comptime

Generate an arbitrary monomorphized AST to be reversed into a valid comptime Noir, with a single comptime function called from main with literal arguments.

make_name 🔒

Derive a variable name from the ID.

program_wrap_expression

Build a program with the single main function returning the result of a given expression (used for conversion of the comptime interpreter execution results for comparison)

Type Aliases

Name 🔒

Name of a variable.