noir_ast_fuzzer::program
Module func

Structs

Flags 🔒

Control what kind of expressions we can generate, depending on the surrounding context.

FunctionContext 🔒

Context used during the generation of a function body.

FunctionDeclaration 🔒

Something akin to a forward declaration of a function, capturing the details required to:

LValueWithMeta 🔒

Helper data structure for generating the lvalue of assignments.

Constants

CONSTRAIN_MSG_TYPE 🔒

Use random strings to identify constraints.

Functions

can_call 🔒

Help avoid infinite recursion by limiting which function can call which other one.

index_name 🔒

Make a name for a local index variable.

local_name 🔒

Make a name for a local variable.

Type Aliases

TrackedExpression 🔒

We need to track whether expressions are coming from dynamic program inputs.