noir_ast_fuzzer::program::func
Struct FunctionContext

pub(super) struct FunctionContext<'a> {

    ctx: &'a mut Context,
    id: FuncId,
    next_local_id: u32,
    next_ident_id: u32,
    budget: usize,
    globals: Scope<GlobalId>,
    locals: ScopeStack<LocalId>,
    dynamics: Stack<HashMap<LocalId, bool>>,
    in_loop: bool,
    in_no_dynamic: bool,
    in_dynamic: bool,
    call_targets: BTreeMap<CallableId, HashSet<Type>>,
    has_call: bool,

}

Context used during the generation of a function body.

Fields
ctx: &'a mut Context

Top level context, to access global variables and other functions.

id: FuncId

Self ID.

next_local_id: u32

Every variable created in the function will have an increasing ID, which does not reset when variables go out of scope.

next_ident_id: u32

Every identifier created in the function will have an increasing ID, which does not reset when variables go out of scope.

budget: usize

Number of statements remaining to be generated in the function.

globals: Scope<GlobalId>

Global variables.

locals: ScopeStack<LocalId>

Variables accumulated during the generation of the function body, initially consisting of the function parameters, then extended by locally defined variables. Block scopes add and remove layers.

dynamics: Stack<HashMap<LocalId, bool>>

Indicate which local variables are derived from function inputs.

in_loop: bool

Indicator of being in a loop (and hence able to generate break and continue statements)

in_no_dynamic: bool

Indicator of computing an expression that should not contain dynamic input.

in_dynamic: bool

Indicator of being affected by dynamic input, in which case we should refrain from using expression that requires no-dynamic mode.

call_targets: BTreeMap<CallableId, HashSet<Type>>

All the functions callable from this one, with the types we can produce from their return value.

has_call: bool

Indicate that we have generated a Call.

Implementations

impl<'a> FunctionContext<'a>

pub(super) fn new(ctx: &'a mut Context, id: FuncId) -> Self

pub(super) fn gen_body(self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate the function body.

pub(super) fn gen_body_with_lit_call( self, u: &mut Unstructured<'_>, callee_id: FuncId, ) -> Result<Expression>

Generate the function body, wrapping a function call with literal arguments. This is used to test comptime functions, which can only take those.

fn decl(&self) -> &FunctionDeclaration

Get the function declaration.

fn unconstrained(&self) -> bool

Is this function unconstrained.

fn is_main(&self) -> bool

Is this the main function?

fn config(&self) -> &Config

fn max_depth(&self) -> usize

The default maximum depth to start from. We use max_depth to limit the complexity of expressions such as binary ones, array indexes, etc.

fn next_local_id(&mut self) -> LocalId

Get and increment the next local ID.

fn next_ident_id(&mut self) -> IdentId

Get and increment the next ident ID.

fn enter_scope(&mut self)

Enter a new local scope.

fn exit_scope(&mut self)

Exit the current local scope.

fn is_dynamic(&self, id: &LocalId) -> bool

Check if a variable is derived from dynamic input.

A variable can become statically known after re-assignment.

fn set_dynamic(&mut self, id: LocalId, is_dynamic: bool)

Mark a variable as dynamic or not dynamic.

fn can_be_used_in_dynamic(&self, src_type: &Type, tgt_type: &Type) -> bool

Check if a source type can be used inside a dynamic input context to produce some target type.

fn choose_producer( &self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<Option<VariableId>>

Choose a producer for a type, preferring local variables over global ones.

fn decrease_budget(&mut self, amount: usize)

Decrease the budget by some amount.

fn get_variable(&self, id: &VariableId) -> &(bool, String, Type)

Get a local or global variable.

Panics if it doesn’t exist.

fn gen_literal( &self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<Expression>

Generate a literal expression of a certain type.

fn gen_expr( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, flags: Flags, ) -> Result<(Expression, bool)>

Generate an expression of a certain type.

While doing so, enter and exit blocks, and add variables declared to the context, so expressions down the line can refer to earlier variables.

This will always succeed, because we can always return a literal expression.

fn gen_expr_from_vars( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Try to generate an expression with a certain type out of the variables in scope.

fn gen_expr_from_source( &mut self, u: &mut Unstructured<'_>, (src_expr, src_dyn): (Expression, bool), src_type: &Type, src_mutable: bool, tgt_type: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Try to generate an expression that produces a target type from a source, e.g. given a source type of [(u32, bool); 4] and a target of u64 it might generate my_var[2].0 as u64.

The src_mutable parameter indicates whether we can take a mutable reference over the source.

Returns None if there is no way to produce the target from the source.

fn gen_vector_access<F>( &mut self, u: &mut Unstructured<'_>, (src_expr, src_dyn): (Expression, bool), src_type: &Type, src_mutable: bool, tgt_type: &Type, max_depth: usize, access_item: F, ) -> Result<Option<(Expression, bool)>>
where F: FnOnce(&mut Self, Ident, Expression) -> Expression,

Generate code to index an arbitrary item in a vector.

Since we don’t know the length of the vector at compile time, this can involve creating a temporary variable, getting the length at runtime, and using modulo to limit some random index to the runtime length.

fn gen_index( &mut self, u: &mut Unstructured<'_>, len: u32, max_depth: usize, ) -> Result<(Expression, bool)>

Generate an arbitrary index for an array.

This can be either a random int literal, or a complex expression that produces an int.

fn gen_unary( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Try to generate a unary expression of a certain type, if it’s amenable to it, otherwise return None.

fn gen_binary( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Try to generate a binary expression of a certain type, if it’s amenable to it, otherwise return None.

fn gen_block( &mut self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<(Expression, bool)>

Generate a block of statements, finally returning a target type.

This should always succeed, as we can always create a literal in the end.

We might use dynamic input in the block and not return a dynamic result from it. For example:

fn main(a: u32) -> u32 {
    let b: u32 = {
        let c = a + 1;
        2
    };
    b
}

fn gen_stmt(&mut self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate a statement, which is an expression that doesn’t return anything, for example loops, variable declarations, etc.

Because the statement doesn’t return anything (return unit) we don’t track whether it used dynamic inputs; there is nothing to propagate.

fn gen_let(&mut self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate a Let statement with arbitrary type and value.

fn let_var( &mut self, mutable: bool, typ: Type, expr: Expression, add_to_scope: bool, is_dynamic: bool, make_name: impl Fn(LocalId) -> String, ) -> Expression

Add a new local variable and return a Let expression.

If add_to_scope is false, the value will not be added to the locals.

fn let_var_and_ident( &mut self, mutable: bool, typ: Type, expr: Expression, add_to_scope: bool, is_dynamic: bool, make_name: impl Fn(LocalId) -> String, ) -> (Expression, Ident)

Add a new local variable and return a Let expression along with an Ident to refer it by.

fn gen_assign(&mut self, u: &mut Unstructured<'_>) -> Result<Option<Expression>>

Assign to a mutable variable, if we have one in scope.

It resets the dynamic flag of the assigned variable.

fn gen_lvalue( &mut self, u: &mut Unstructured<'_>, lvalue: LValue, typ: Type, ) -> Result<LValueWithMeta>

Generate an lvalue to assign to a local variable, or some part of it, if it’s a compound type.

Say we have an array: a: [[u32; 2]; 3]; we call it with a, and it might return a, a[i], or a[i][j].

fn gen_print(&mut self, u: &mut Unstructured<'_>) -> Result<Option<Expression>>

Generate a println statement, if there is some printable local variable.

For now this only works in unconstrained code. For constrained code we will need to generate a proxy function, which we can do as a follow-up pass, as it has to be done once per function signature.

fn gen_constrain( &mut self, u: &mut Unstructured<'_>, ) -> Result<Option<Expression>>

Generate a constrain statement, if there is some local variable we can do it on.

Arbitrary constraints are very likely to fail, so we don’t want too many of them, otherwise they might mask disagreements in return values.

fn gen_if( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, flags: Flags, ) -> Result<(Expression, bool)>

Generate an if-then-else statement or expression.

fn gen_for(&mut self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate a for loop.

fn gen_call( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Generate a function call to any function in the global context except main, if the function returns the target type, or something we can use to produce that type.

Whether a call is dynamic depends on whether it has dynamic arguments, and whether we are crossing an ACIR-to-Brillig boundary.

fn gen_lit_call( &mut self, u: &mut Unstructured<'_>, callee_id: FuncId, ) -> Result<Expression>

Generate a call to a specific function, with arbitrary literals for arguments (useful for generating comptime wrapper calls)

fn gen_loop(&mut self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate a loop loop.

fn gen_while(&mut self, u: &mut Unstructured<'_>) -> Result<Expression>

Generate a while loop.

fn gen_loop_size(&self, u: &mut Unstructured<'_>) -> Result<usize>

Choose a random maximum guard size for loop and while to match the average of the size of a for.

fn gen_match( &mut self, u: &mut Unstructured<'_>, typ: &Type, max_depth: usize, ) -> Result<Option<(Expression, bool)>>

Generate a match expression, returning a given type.

Match needs a variable; if we don’t have one to produce the target type from, it returns None.

fn gen_num_field( &mut self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<SignedField>

Generate a random field that can be used in the match constructor of a numeric type.

fn gen_num_match_constructor( &mut self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<Constructor>

Generate a match constructor for a numeric type.

fn gen_guaranteed_call_from_main( &mut self, u: &mut Unstructured<'_>, ) -> Result<Option<Expression>>

If this is main, and we could have made a call to another function, but we didn’t, ensure we do, so as not to let all the others we generate go to waste.

fn find_global_function_with_signature( &mut self, u: &mut Unstructured<'_>, typ: &Type, ) -> Result<Expression>

Find a global function matching a type signature.

For local functions we use gen_expr_from_vars.

fn callable_expr(&mut self, callee_id: CallableId) -> Expression

Expression to use in a Call for the function (pointer).

If the ID is something like f: &mut &mut fn(...) -> ... then it will return (*(*f)).

fn func_ident(&mut self, callee_id: FuncId) -> Ident

Identifier for a global function.

fn local_ident(&mut self, id: LocalId) -> Ident

Identifier for a local variable.

fn local_type(&self, id: LocalId) -> &Type

Type of a local variable.

fn next_loop_index(&mut self) -> (LocalId, String, Ident)

Create a loop index variable.

fn callable_signature(&self, callee_id: CallableId) -> (Vec<Type>, Type, bool)

Get the parameter types and return type of a callable function.

fn indirect_ref_mut( &mut self, (expr, is_dyn): (Expression, bool), typ: Type, ) -> Expression

Create a block with a let binding, then return a mutable reference to it. This is used as a workaround when we need a mutable reference over an immutable value.

fn indirect_global( &mut self, id: GlobalId, mutable: bool, add_to_scope: bool, ) -> (LocalId, String, Expression)

Create a local let binding over a global variable.

Returns the local ID and the Let expression.

fn indirect_local( &mut self, id: LocalId, mutable: bool, add_to_scope: bool, ) -> (LocalId, String, Expression)

Create a local let binding over a local variable.

Returns the local ID and the Let expression.

fn call_array_len( &mut self, array_or_vector: Expression, typ: Type, ) -> Expression

Construct a Call to the array_len builtin function, calling it with the identifier of a vector or an array.

fn call_vector_builtin( &mut self, name: &str, return_type: Type, arg_types: Vec<Type>, args: Vec<Expression>, ) -> Expression

Construct a Call to one of the vector_* builtin functions.

fn call_vector_push( &mut self, vector_type: Type, item_type: Type, vector: Expression, is_front: bool, item: Expression, ) -> Expression

Construct a Call to the vector_push_front or vector_push_back builtin function.

fn call_vector_pop( &mut self, vector_type: Type, item_type: Type, vector: Expression, is_front: bool, ) -> Expression

Construct a Call to the vector_pop_front or vector_pop_back builtin function.

fn call_vector_remove( &mut self, vector_type: Type, item_type: Type, vector: Expression, idx: Expression, ) -> Expression

Construct a Call to the vector_remove builtin function.

fn call_vector_insert( &mut self, vector_type: Type, item_type: Type, vector: Expression, idx: Expression, item: Expression, ) -> Expression

Construct a Call to the vector_insert builtin function.

fn avoid_index_out_of_bounds(&self, u: &mut Unstructured<'_>) -> Result<bool>

Random decision whether to allow “Index out of bounds” errors to happen on a specific array or vector access operation.

If Config::avoid_index_out_of_bounds is turned on, then this is always true.

It also returns true when in_no_dynamic mode is on, because an overflowing index might not be simplified out of the SSA in ACIR, and end up being considered a dynamic index, and leave reference allocations until ACIR gen, where they fail.