noir_ast_fuzzer::program::rewrite

Module limit

Structs

LimitContext 🔒

Constants

LIMIT_NAME 🔒

Functions

add_recursion_limit 🔒

To avoid the potential of infinite recursion at runtime, add a ctx_limit: &mut u32 parameter to all functions, which we use to limit the number of recursive calls.

ctx_limit_type_for_func_param 🔒

Decide how to pass the recursion limit to function: by value or by ref.

modify_function_pointer_param_type 🔒

Recursively modify function pointers in the param type.

modify_function_pointer_param_types 🔒

Go through the types of each function parameter. If they are function pointers, then they need the context, depending on the callee type.

modify_function_pointer_param_values 🔒

Go through the call arguments and update global function pointers to their proxy equivalents if necessary.