noir_ast_fuzzer::program
Module types

Constants

U8

U32

Functions

can_be_global

We can only use globals that can be evaluated at comptime. Some types don’t compile in Noir, so avoid those as we couldn’t put any related failures into an integration test.

can_be_main

Check if a type can be used in the main function.

can_be_matched

Check if a variable with a given type can be used in a match.

can_binary_op_err_by_zero

Can the binary operation fail because the RHS is zero. These operations can fail because of an unfortunate combination of expressions, such as a / (a % a) or a % (a % a), but it’s less like to occur than the overflows.

can_binary_op_overflow

Can the binary operation result in an overflow. These are operations that commonly fail with random constants.

can_binary_op_return

Check if a certain binary operation can return a target type as output.

can_binary_op_return_from_input

Check if a certain binary operation can take a type as input and produce the target output.

can_binary_return

Can the type be returned by some BinaryOp.

can_unary_return

Can the type be returned by some UnaryOp.

contains_reference

Check if the type contains any references.

contains_vector

Check if the type contains any references.

is_array_or_vector 🔒

Check if the type is a vector or an array.

is_bool

Check if the type works with UnaryOp::Not

is_function 🔒

Check if the type is a function.

is_numeric

Check if the type is a number.

is_printable

Check if the type can be used with a println statement.

is_reference

Check if the type is a reference wrapping another.

is_unit

Check if a type is Unit.

ref_mut

Reference an expression into a target type

to_hir_type

Convert the type back into a HIR equivalent (not necessarily the original HIR type).

type_depth

Calculate the depth of a type.

types_produced

Collect all the sub-types produced by a type.

unref 🔒

Peel off all reference types, to get to a concrete underlying type.

unref_mut 🔒

Peel off all reference types, to get to a concrete underlying type.