noir_ast_fuzzer
Struct Config

pub struct Config {

    pub max_globals: usize,
    pub min_functions: usize,
    pub max_functions: usize,
    pub max_function_args: usize,
    pub max_function_size: usize,
    pub max_block_size: usize,
    pub max_depth: usize,
    pub max_tuple_size: usize,
    pub max_array_size: usize,
    pub max_loop_size: usize,
    pub vary_loop_size: bool,
    pub max_recursive_calls: usize,
    pub max_match_cases: usize,
    pub expr_freqs: Freqs,
    pub stmt_freqs_acir: Freqs,
    pub stmt_freqs_brillig: Freqs,
    pub force_brillig: bool,
    pub avoid_overflow: bool,
    pub avoid_index_out_of_bounds: bool,
    pub avoid_err_by_zero: bool,
    pub avoid_negative_int_literals: bool,
    pub avoid_large_int_literals: bool,
    pub avoid_loop_control: bool,
    pub avoid_lambdas: bool,
    pub avoid_print: bool,
    pub avoid_constrain: bool,
    pub avoid_match: bool,
    pub avoid_vectors: bool,
    pub comptime_friendly: bool,

}

AST generation configuration.

Fields
max_globals: usize

Maximum number of global definitions.

min_functions: usize

Minimum number of functions (other than main) to generate.

max_functions: usize

Maximum number of functions (other than main) to generate.

max_function_args: usize

Maximum number of arguments a function can have.

max_function_size: usize

Maximum number of statements to aim for in a function body.

max_block_size: usize

Maximum number of statements in a block.

max_depth: usize

Maximum nesting depth for complex expressions.

max_tuple_size: usize

Maximum number of fields for tuples.

max_array_size: usize

Maximum size for arrays.

max_loop_size: usize

Maximum size of for loop ranges, which affects unrolling in ACIR.

vary_loop_size: bool

Whether to choose the backstop for loop and while randomly.

max_recursive_calls: usize

Maximum number of recursive calls to make at runtime.

max_match_cases: usize

Maximum number of match cases.

expr_freqs: Freqs

Frequency of expressions, which produce a value.

stmt_freqs_acir: Freqs

Frequency of statements in ACIR functions.

stmt_freqs_brillig: Freqs

Frequency of statements in Brillig functions.

force_brillig: bool

Whether to force all functions to be unconstrained.

avoid_overflow: bool

Try to avoid overflowing operations. Useful when testing the minimal pipeline, to avoid trivial failures due to multiplying or adding constants.

avoid_index_out_of_bounds: bool

Try to avoid โIndex out of boundsโ by using modulo to limit indexing to the range that an array or vector is expected to contain.

This is easy to trigger (a random u32 will most certainly be out of the range of the arrays we generate), so by default it is โonโ. When itโs โoffโ, a random decision is taken for each index operation whether to apply the modulo or not.

avoid_err_by_zero: bool

Try to avoid operations that can result in error when zero is on the RHS.

avoid_negative_int_literals: bool

Avoid using negative integer literals where the frontend expects unsigned types.

avoid_large_int_literals: bool

Avoid using large integer literals where the frontend expects 32 bits.

avoid_loop_control: bool

Avoid using loop control (break/continue).

avoid_lambdas: bool

Avoid using function pointers in parameters.

avoid_print: bool

Avoid print statements.

avoid_constrain: bool

Avoid using constrain statements.

avoid_match: bool

Avoid match statements and expressions.

avoid_vectors: bool

Avoid using the vector type.

comptime_friendly: bool

Only use comptime friendly expressions.