Crate noir_ast_fuzzer

Re-exports

pub use compare::input_value_to_ssa;

pub use compare::input_values_to_ssa;

Modules

abi 🔒

compare

expr

input 🔒

program 🔒

Module responsible for generating arbitrary [Program] ASTs.

rewrite

scope

types

Structs

Config

AST generation configuration.

DisplayAstAsNoir

Wrapper around Program that prints the AST as close to being able to copy-paste it as a Noir program as we can get.

DisplayAstAsNoirComptime

Wrapper around Program that prints its AST as close to Noir syntax as we can get, making it comptime. The AST must be specifically prepared to include a main function consisting of a comptime wrapped call to a comptime (or unconstrained) marked function.

Functions

arb_inputs

Generate arbitrary inputs for a compiled program according to the ABI.

arb_program

Generate an arbitrary monomorphized AST.

arb_program_comptime

Generate an arbitrary monomorphized AST to be reversed into a valid comptime Noir, with a single comptime function called from main with literal arguments.

program_abi

Generate the [Abi] interface of a [Program].

program_wrap_expression

Build a program with the single main function returning the result of a given expression (used for conversion of the comptime interpreter execution results for comparison)