noir_ast_fuzzer
Module compare

Modules

compiled 🔒

Compare an arbitrary AST compiled into bytecode and executed with the VM.

comptime 🔒

Compare an arbitrary AST executed as Noir with the comptime interpreter vs compiled into bytecode and ran through a VM.

interpreted 🔒

Compare an arbitrary AST compiled into SSA and executed with the SSA interpreter at some stage of the SSA pipeline.

logging 🔒

We can turn on the logging of artifacts by setting the RUST_LOG=debug env var.

Structs

CompareArtifact

CompareCompiled

Compare the execution of equivalent programs, compiled in different ways.

CompareComptime

Compare the execution of a Noir program in pure comptime (via interpreter) vs normal SSA execution.

CompareInterpreted

Inputs for comparing the interpretation of two SSA states of an arbitrary program.

CompareOptions

Subset of [SsaEvaluatorOptions] that we want to vary.

ComparePass

The state of the SSA after a particular pass in the pipeline.

FailedOutput

Error returned from the circuit, with tracked side effects.

PassedOutput

Optional return value with the tracked side effects.

Enums

CompareResult

Possible outcomes of the differential execution of two equivalent programs.

Traits

Comparable

Help ignore results and errors we find equivalent between executions.

HasPrograms

Help iterate over the program(s) in the comparable artifact.

Functions

input_value_to_ssa

Convert one ABI encoded input to what the SSA interpreter expects.

input_values_to_ssa

Convert the ABI encoded inputs to what the SSA interpreter expects.

Type Aliases

CompareCompiledResult

The result of the execution of compiled programs, decoded by their ABI.

CompareInterpretedResult

The result of the SSA interpreter execution.

CompareMorph

Compare two equivalent variants of the same program, compiled the same way.

ComparePipelines

Compare the execution the same program compiled in two different ways.