noir_ast_fuzzer
Module input

Modules

dictionary 🔒

int 🔒

uint 🔒

Functions

arb_input_map 🔒

Given the Abi description of a Noir program, generate random InputValues for each circuit parameter.

arb_inputs

Generate arbitrary inputs for a compiled program according to the ABI.

arb_inputs_from_ssa 🔒

Generate arbitrary inputs for an SSA of a program, according to the ABI.

arb_value_from_abi_type 🔒

Create a strategy for generating random values for an AbiType.

arb_value_tree 🔒

Generate a seed and use it to generate an arbitrary value using the proptest strategy.