noir_ast_fuzzer::program
Module expr

Functions

assign_ident

Assign a value to an identifier.

assign_ref

Assign a value to a mutable reference.

binary

Make a binary expression.

cast

Cast an expression to a target type.

deref

Dereference an expression into a target type

equal

Make an == expression.

exists

Check if an Expression or any of its descendants match a predicate.

extend_block

Append statements to a given block.

gen_literal

Generate a literal expression according to a type.

gen_range

Generate a literals for loop ranges with signed/unsigned integers with bits 8, 16, 32 or 64 bits, in a way that start is not higher than the end, and the maximum difference between them is limited, so that we don’t get huge unrolled loops.

has_call

Check if an Expression contains any Call another function, in any of its descendants. Calls made to oracles such as println don’t count.

ident 🔒

Make an Ident expression out of a variable.

ident_inner 🔒

Make an Ident out of a variable.

if_else

Make an if/else expression.

if_then

Create an if expression, with an optional else.

index_modulo

Take an integer expression and make sure it fits in an expected len by taking a modulo.

int_literal

Integer literal, can be positive or negative depending on type.

is_deref 🔒

Is the expression dereferencing something.

is_immutable_ident 🔒

Is the expression an identifier of an immutable variable

let_var

Create a variable.

lit_bool

Boolean literal.

modulo

Make a modulo expression.

prepend

Prepend an expression to a destination.

prepend_block

Prepend statements to a given block.

range_modulo

Take an integer expression and make sure it’s no larger than max_size.

reachable_functions

Collect all the functions referred to by their ID in the expression and its descendants.

ref_mut

Reference an expression as a target type

ref_with_mut 🔒

replace

Replace an expression with another one, passing its current value to a function.

u8_literal

8-bit unsigned int literal, used in bit shifts.

u32_literal

32-bit unsigned int literal, used in indexing arrays.

unary

Make a unary expression.

unref_mut 🔒

Peel back any dereference operators until we get to some other kind of expression.